requirement.txt
APScheduler==3.6.3
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.10
mysqlclient==2.0.1
numpy==1.19.1
pandas==1.0.5
pyactiveresource==2.2.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ShopifyAPI==8.0.2
six==1.15.0
sqlparse==0.3.1
tzlocal==2.1
urllib3==1.25.9
XlsxWriter==1.2.9
